Moved recoveryKey test into a seaparate file with proper assertion in place
