+++ src/IO.Ably.Tests.Shared/Realtime/ConnectionSpecs/ConnectionRecoverySpecs.cs
@@ -65,25 +65,6 @@ namespace IO.Ably.Tests.Realtime.ConnectionSpecs
             client.Connection.RecoveryKey.Should().Be(expectedRecoveryKey);
         }
 
-        [Fact(Skip = "Test independently passes, but last assertion (Connection.MessageSerial = 45) fails when whole test file is run")]
-        [Trait("spec", "RTN16i")]
-        [Trait("spec", "RTN16f")]
-        [Trait("spec", "RTN16j")]
-        public async Task RecoveryKey_MsgSerialShouldNotBeSentToAblyButShouldBeSetOnConnection()
-        {
-            var recoveryKey =
-                "{\"connectionKey\":\"uniqueKey\",\"msgSerial\":45,\"channelSerials\":{\"channel1\":\"1\",\"channel2\":\"2\",\"channel3\":\"3\"}}";
-            var client = GetClientWithFakeTransport(options => { options.Recover = recoveryKey; });
-
-            var transportParams = await client.ConnectionManager.CreateTransportParameters("https://realtime.ably.io");
-            var paramsDict = transportParams.GetParams();
-            paramsDict.ContainsKey("recover").Should().BeTrue();
-            paramsDict["recover"].Should().Be("uniqueKey");
-            paramsDict.ContainsKey("msg_serial").Should().BeFalse();
-
-            client.Connection.MessageSerial.Should().Be(45);
-        }
-
         public ConnectionRecoverySpecs(ITestOutputHelper output)
             : base(output)
         {
